Add plots for infeasible objective set to 0
Plots showing the trend of the objective over the number of alternatives
are central to our work. Before, we only showed aggregate (box or median)
quality that automatically excluded infeasible feature sets (as their
quality is NaN). However, it also makes sense to consider the objective of
infeasible feature sets to be 0 (i.e., like no feature has been selected,
at least for the MI selector we use for plotting). Thus, we add such plots
to complement the existing plots.
We set k=10 in the num-alternatives plot for consistency to the tau plots.
We use mean instead of median in the tau plots for a more informative view
(in "fillna" case, median soon becomes 0 and lines therefore constant
rather than showing a stedy decrease).