Update Evaluation 6.1-6.3 (datasets, Q, FS) with 2022-12-01 run
Integrate mRMR and simultaneous search with min-objective (though the
latter is not explictly discussed in these sections). Update some numbers
in the text and the figures. Most results are very similar to the old ones,
but greedy-wrapper performance seems to differ a lot. This might be related
to the solver change (CbC -> SCIP). In the old results, feature sets found
by the wrapper often seemed to involve high feature indices (I only
eyeballed this quickly). In the new results, feature sets found by the
wrapper often seem to involve low feature indices. I guess this is related
to the order in which the solver swaps features. Maybe the features with
higher indices tend to be more important for our datasets.