Double down on problem size for NP-proof
In bb5c139c30, we changed on NP proof to assume all constraints were part
of the problem input. Actually, it is more reasonable to assume that
observing constraints is part of the solution algorithm, while we only
need to provide the parameter values. Thus, we provide both these
interpretations now (lead to same result anyway).