Move most complexity stuff into Approach 1/2 (delete)
Apart from adapting cross-references and section intros, we largely left
the text as-is. Cuts are necessary. Since Git diff might become confused
by the large amount of text moved and therefore show also other parts of
the text as changed, we use two commit. This commit deletes text from the
Appendix, the next commit inserts subsubsections into the Approach. The
currrent commit should not be used on its own since other references in the
text are already adapted to the new section structure.