Adapt labels for univariate heuristics and add To Dos
While the approach of heuristics is described already, we need to integrate
them into intro, conclusion, experimental design, and evaluation. Apart
from minor changes (e.g., change "univariate objective" to "univariate
feature qualities"), we left the text as-is for now but added comments
where changes will be necessary. Also, we adapted labels for the heuristics
sections to refer to "approach" instead of "appendix".